!! FILE RESTRUCTURING + documentation
