+++ graph_creation.py
@@ -1,37 +1,30 @@
+'''
+The graph is being created here, and then being imported in the mainapp file. 
+This ensure that mainapp.py is not bloated with code
+The graph creation is rather standard for langgraph
+All tools, routers and agents prompts are created in other files
+'''
+
 import os
-from langgraph.graph import StateGraph, END
-from langchain_core.messages import (
-    HumanMessage,
-)
-from langgraph.prebuilt import ToolNode
-from langgraph.graph import END, StateGraph, START
 import functools
+
+#langgraph imports
+from langgraph.graph import StateGraph, END, START
+from langgraph.prebuilt import ToolNode
 from langchain_openai import ChatOpenAI
-from langgraph.checkpoint import *
+
+# imports from other files
 from routers import *
 from agents import *
 from tools import *
-
-import logging
-from fastapi import FastAPI, HTTPException, BackgroundTasks
-from pydantic import BaseModel
-from fastapi.responses import JSONResponse
-import asyncio
-
-from state import message_state
-
-app = FastAPI()
 
 os.environ["LANGSMITH_API_KEY"] = "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"
 os.environ["LANGCHAIN_TRACING_V2"] = "true"
 os.environ["LANGCHAIN_PROJECT"] = "tutor agents"
 os.environ["OPENAI_API_KEY"] = "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"
+
 GPT_MODEL = "gpt-4o"
 llm = ChatOpenAI(model=GPT_MODEL)
-
-global_prompts_list =[]
-agent_activation_order = []
-userID = '001'
 
 communicator_agent = create_agent(
     agentName='communicator', 
@@ -199,7 +192,6 @@ workflow.add_conditional_edges(
 )
 
 workflow.add_conditional_edges(
-    #! for mutliple agents: USE LAMBDA EXPRESSION TO ALWAYS SEND MSG BACK TO SENDER + no need to send back agent type in get prompt
     "tutor_call_tool",
     route_back_to_tutor,
     {'conversational': 'conversational', 
@@ -211,74 +203,4 @@ workflow.add_conditional_edges(
 
 workflow.add_edge(START, "communicator")
 
-graph = workflow.compile() #checkpointer=memory
-
-# Configure logging
-logging.basicConfig(level=logging.INFO,
-                    format='%(asctime)s - %(levelname)s - %(message)s')
-
-# Create a logger object
-logger = logging.getLogger(__name__)
-
-class ConversationRequest(BaseModel):
-    startBool: bool
-    userID: str
-
-class UserInputRequest(BaseModel):
-    content: str
-
-class AcknowledgmentRequest(BaseModel):
-    ack: bool
-
-async def continue_graph_execution(messages):
-    async for s in graph.astream({"messages": messages}, {"recursion_limit": 100}):
-        if "__end__" not in s:
-            print(s)
-            print("----")
-            # Update the shared state with the current content and agent name
-            await asyncio.sleep(0)  # Yield control to the event loop
-
-
-@app.post("/startConversation")
-async def begin_graph_stream(request: ConversationRequest, background_tasks: BackgroundTasks):
-    if request.startBool:
-        try:
-            messages = [
-                HumanMessage(content=f"Communicator, the user ID is {request.userID}. Please start with your task.")
-            ]
-            background_tasks.add_task(continue_graph_execution, messages)
-            return JSONResponse(content={"message": "Conversation started"}, status_code=200)
-        except Exception as e:
-            return JSONResponse(content={"error": str(e)}, status_code=500)
-    else:
-        raise HTTPException(status_code=400, detail="startBool must be true")
-
-@app.get("/getAIMessage")
-async def get_ai_message():
-    print('entering get function...')
-    message_data = await message_state.wait_for_update()
-    response_payload = {
-        "agent_name": message_data["agent_name"],
-        "content": message_data["content"]
-    }
-    print('exiting get function...')
-    return JSONResponse(content=response_payload, status_code=200)
-
-@app.post("/userInput")
-async def receive_user_input(request: UserInputRequest):
-    try:
-        message_state.update_user_input(request.content)
-        return JSONResponse(content={"message": "User input received"}, status_code=200)
-    except Exception as e:
-        return JSONResponse(content={"error": str(e)}, status_code=500)
-
-@app.post("/acknowledgeMessage")
-async def acknowledge_message(request: AcknowledgmentRequest):
-    try:
-        if request.ack:
-            message_state.acknowledge_message()
-            return JSONResponse(content={"message": "Acknowledgment received"}, status_code=200)
-        else:
-            raise HTTPException(status_code=400, detail="Acknowledgment must be true")
-    except Exception as e:
-        return JSONResponse(content={"error": str(e)}, status_code=500)
+graph = workflow.compile()
