+++ routers.py
@@ -1,5 +1,8 @@
+'''
+Routers used to decide what path to follow between agent and tool nodes
+TODO: add more documentation to the routers, mostly rename variable for more clarity
+'''
 
-import time
 from typing import Literal
 from state import message_state
 
@@ -22,7 +25,6 @@ def route_to_tutor(state) -> Literal["conversational", "no_more_lesson"]:
 
 async def router_tutor(state) -> Literal["call_tool", "continue", "FINAL REPORT"]:
     ''' BUG: sometimes the tutor starts the convo without calling its tool, fix asap @urgent'''
-    #? create seperate router for each tutor agent or generalise it? -should be able to generalize
     print('-- tutor agent router --')
     messages = state["messages"]
     last_message = messages[-1]
@@ -31,18 +33,18 @@ async def router_tutor(state) -> Literal["call_tool", "continue", "FINAL REPORT"
     if "REPORT DONE" in last_message.content:
         return "FINAL REPORT"
     elif "USER TURN":
-        #! So this print statement here should be what is being returned to the front end !!
+        # DEBUGGING
         print('AI ASSISTANT: ', last_message.content)
         
+        # THIS IS TO UPDATE THE LAST MESSAGE TO THEN SEND IT TO THE CLIENT
         message_state.update_content(last_message.content, last_message.name)
-
+        # WAIT FOR ACK TO SEE IF CLIENT RECIEVED AIMESSAGE 
         await message_state.wait_for_acknowledgment()
 
         print('You:')
-        #! and this here should be a get request to the front end !!
+        # WAIT FOR THE CLIENT TO SEND THE USER INPUT
         user_input = await message_state.wait_for_input()
-        #print('YOU: ', user_input)
-        #add this to messages
+      
         last_message.content += " user response: " + user_input
         return "continue" 
     return "call_tool" #forcing call tool (sometimes agent starts conversation anyways, need to fix that)
@@ -62,7 +64,7 @@ def route_back_to_tutor(state) -> Literal['conversational', 'reader', 'listening
         return 'questionAnswering'
     if 'FINAL REPORT' and 'grammarSummary' in last_message.content:
         return 'grammarSummary'
-    if 'conversational' in last_message.content: #? might have to change for multiple tutors
+    if 'conversational' in last_message.content: 
         return 'conversational'
     if 'reader' in last_message.content:
         return 'reader'
@@ -77,10 +79,9 @@ def orchestrator_router(state) -> Literal['continue_to_tracker', 'call_tool']:
     print('-- orchestrator router --')
     messages = state['messages']
     last_message = messages[-1]
-    if last_message.tool_calls: #otherwise call your tool
+    if last_message.tool_calls: 
         return 'call_tool'
     if 'continue' in last_message.content: #if both tools have been called go to the tracker
-        # clear_memory(memory=memory, thread_id="123456", thread_ts="123456")
         return 'continue_to_tracker'
     
 async def communicator_router(state) -> Literal['continue', 'go_orchestrator', 'call_tool']:
@@ -91,17 +92,16 @@ async def communicator_router(state) -> Literal['continue', 'go_orchestrator', '
         return "call_tool"
     if 'go_orchestrator' in last_message.content:
         return 'go_orchestrator'
-    #! So this print statement here should be what i
     print('AI ASSISTANT: ', last_message.content)
 
+    # THIS IS TO UPDATE THE LAST MESSAGE TO THEN SEND IT TO THE CLIENT
     message_state.update_content(last_message.content, last_message.name)
-
+    # WAIT FOR ACK TO SEE IF CLIENT RECIEVED AIMESSAGE
     await message_state.wait_for_acknowledgment()
 
     print('You:')
-    #! and this here should be a get request to the front end !!
+    # WAIT FOR THE CLIENT TO SEND THE USER INPUT
     user_input = await message_state.wait_for_input()
-    # print('YOU: ', user_input)
-    #add this to messages
+    
     last_message.content += " user response: " + user_input
     return 'continue'
